Reverted some previous changes...

+++ src/raylib.h
@@ -423,8 +423,8 @@ typedef struct AudioStream {
     rAudioBuffer *buffer;           // Pointer to internal data used by the audio system
 
     unsigned int sampleRate;        // Frequency (samples per second)
-    unsigned short sampleSize;        // Bit depth (bits per sample): 8, 16, 32 (24 not supported)
-    unsigned short channels;          // Number of channels (1-mono, 2-stereo)
+    unsigned int sampleSize;        // Bit depth (bits per sample): 8, 16, 32 (24 not supported)
+    unsigned int channels;          // Number of channels (1-mono, 2-stereo)
 } AudioStream;
 
 // Sound source type
